Updated Navbar and created files for updated Navbar routes
