+++ pages/member.js
@@ -1,0 +1,27 @@
+/* eslint-disable react-hooks/exhaustive-deps */
+import React, { useEffect, useState } from 'react';
+import { getMember } from '../api/memberData';
+import { useAuth } from '../utils/context/authContext';
+import MemberCard from '../components/MemberCard';
+
+export default function Team() {
+  const [members, setMembers] = useState([]);
+
+  const { user } = useAuth();
+
+  const getAllMembers = () => {
+    getMember(user.uid).then(setMembers);
+  };
+
+  useEffect(() => {
+    getAllMembers();
+  }, []);
+
+  return (
+    <div className="d-flex flex-wrap justify-content-between">
+      {members.map((member) => (
+        <MemberCard key={member.firebaseKey} memberObj={member} onUpdate={getAllMembers} />
+      ))}
+    </div>
+  );
+}
