+++ pages/team/[firebaseKey].js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function ViewTeam() {
+  return (
+    <h1>View Members in Team</h1>
+  );
+}
