+++ pages/team/new.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function AddTeam() {
+  return (
+    <h1>
+      Add Team
+    </h1>
+  );
+}
